rgw/requirements.txt
# requires python3 
boto
boto3
names
PyYaml
ConfigParser
python-swiftclient
swiftly
simplejson
# for v1
psutil